+++ frontend/src/pages/xray/outbounds/outbounds-tab-types.ts
@@ -1,0 +1,7 @@
+export interface OutboundRow {
+  key: number;
+  tag?: string;
+  protocol?: string;
+  streamSettings?: { network?: string; security?: string };
+  settings?: Record<string, unknown>;
+}
